+++ src/app/core/crud/crud.controller.ts
@@ -1,23 +1,24 @@
-import { Get, Post, Delete, Body, Param, Patch, HttpStatus } from '@nestjs/common';
+import { Get, Post, Delete, Body, Param, Patch, HttpStatus, Query } from '@nestjs/common';
 import { ApiOperation, ApiResponse } from '@nestjs/swagger';
 import { DeleteResult, UpdateResult } from 'typeorm';
+import { BaseEntitySearchDto } from '@root/app/common/base/base-search.dto';
 import { ICrudService } from './crud.service.model';
 
 export class CrudController<T> {
   constructor(private readonly crudService: ICrudService<T>) {}
 
+  @ApiOperation({ summary: 'Search paginated' })
+  @ApiResponse({ status: HttpStatus.OK, description: 'Found records' })
   @Get()
-  @ApiOperation({ summary: 'List all records.' })
-  @ApiResponse({ status: HttpStatus.OK, description: 'Ok' })
-  async findAll(): Promise<T[]> {
-    return this.crudService.getAll();
+  async search(@Query() options?: BaseEntitySearchDto<T>): Promise<{ items: T[]; total: number }> {
+    return this.crudService.search(options);
   }
 
   @Get(':id')
   @ApiOperation({ summary: 'Retrieve a record by Id.' })
   @ApiResponse({ status: HttpStatus.OK, description: 'Entity retrieved successfully.' })
   @ApiResponse({ status: HttpStatus.NOT_FOUND, description: 'Entity does not exist' })
-  async findById(@Param('id') id: string): Promise<T> {
+  async findById(@Param('id') id: string, @Query() query?: any): Promise<Partial<T>> {
     return this.crudService.findById(id);
   }