+++ src/app/core/crud/crud.service.ts
@@ -7,15 +7,37 @@ import {
   UpdateResult,
   FindOneOptions,
   FindConditions,
+  SelectQueryBuilder,
 } from 'typeorm';
-import { ErrorMessage } from '../../common/enum/error-message.enum';
+import { ErrorMessage } from '@root/app/common/enum/error-message.enum';
 import { ICrudService } from './crud.service.model';
+import { BaseEntitySearchDto } from '@root/app/common/base/base-search.dto';
+import { IRelation } from '@root/app/common/base/relation.interface';
 
 @Injectable()
 export class CrudService<T extends BaseEntity> implements ICrudService<T> {
   private readonly entityName = this.genericRepository.metadata.targetName;
 
   constructor(private readonly genericRepository: Repository<T>) {}
+
+  public async search(options: BaseEntitySearchDto<T>) {
+    const { limit, offset } = options;
+    const alias = this.genericRepository.metadata.targetName;
+    let qb = this.genericRepository.createQueryBuilder(alias);
+
+    if (options.relations.length) {
+      for (const relation of options.relations) {
+        this.joinRelation(qb, alias, relation);
+      }
+    }
+
+    if (options.selectFields.length) {
+      qb = qb.select(options.selectFields.map((col) => `${alias}.${col}`));
+    }
+
+    const [items = [], total = 0] = await qb['limit'](limit)['offset'](offset).getManyAndCount();
+    return { items, total };
+  }
 
   async getAll(): Promise<T[]> {
     return await this.genericRepository.find();
@@ -56,5 +78,17 @@ export class CrudService<T extends BaseEntity> implements ICrudService<T> {
       }
     }
     throw new NotFoundException(ErrorMessage.Common.EntityNotFound(this.entityName));
+  }
+
+  private joinRelation(qb: SelectQueryBuilder<T>, parentEntity: string, relation: string | IRelation<T>) {
+    const config: IRelation<T> = typeof relation === 'string' ? { name: relation, innerJoin: false } : relation;
+    const propertyPath = config.name.includes('.') ? config.name : `${parentEntity}.${config.name}`;
+    const relationAlias = config.name.includes('.') ? config.name.split('.').pop() : config.name;
+    if (config.innerJoin) {
+      qb = qb.innerJoinAndSelect(propertyPath, relationAlias);
+    } else {
+      qb = qb.leftJoinAndSelect(propertyPath, relationAlias);
+    }
+    return qb;
   }
 }