+++ src/app/game/game.controller.ts
@@ -1,15 +1,31 @@
-import { Body, Controller, Param, Patch, Post } from '@nestjs/common';
+import { Body, Controller, Get, Param, Patch, Post, Query } from '@nestjs/common';
 import { ApiOperation, ApiTags } from '@nestjs/swagger';
 import { CreateGameDto, GameDto, UpdateGameDto } from '@root/models/game/game.dto';
 import { CrudController } from '@root/app/core/crud/crud.controller';
 import { GameService } from './game.service';
 import { UpdateResult } from 'typeorm';
+import { GameSearchDto } from '@root/models/game/game-search.dto';
 
 @Controller('games')
 @ApiTags('Game')
 export class GameController extends CrudController<GameDto> {
   constructor(private readonly gameService: GameService) {
     super(gameService);
+  }
+
+  @ApiOperation({ summary: 'Search paginated' })
+  @Get()
+  async search(@Query() query?: GameSearchDto): Promise<{ items: GameDto[]; total: number }> {
+    return this.gameService.search(query);
+  }
+
+  @Get(':id')
+  @ApiOperation({ summary: 'Retrieve a record by Id.' })
+  async findById(@Param('id') id: string, @Query('publisherOnly') publisherOnly?: boolean): Promise<Partial<GameDto>> {
+    if (publisherOnly) {
+      return this.gameService.getPublisherByGameId(id);
+    }
+    return this.gameService.findById(id);
   }
 
   @ApiOperation({ summary: 'Create a new record' })